+++ tests/testthat/test_MsqlBackend.R
@@ -39,6 +39,20 @@ test_that("peaksData,MsqlBackend works", {
     idx <- c(2L, 5L, 1L, 2L, 5L)
     res <- mm8_be[idx]
     expect_identical(peaksData(res), peaksData(mm8_sps@backend[idx]))
+
+    res <- peaksData(mm8_be)
+    res_2 <- peaksData(mm8_be, "mz")
+    expect_true(colnames(res_2[[1]]) == "mz")
+    expect_equal(res[[1]][, "mz"], res_2[[1]][, "mz"])
+
+    res_2 <- peaksData(mm8_be, c("intensity", "mz"))
+    expect_true(all(colnames(res_2[[1]]) == c("intensity", "mz")))
+    expect_equal(res[[1]][, "mz"], res_2[[1]][, "mz"])
+    expect_equal(res[[1]][, "intensity"], res_2[[1]][, "intensity"])
+})
+
+test_that("peaksVariables,MsqlBackend works", {
+    expect_equal(peaksVariables(mm8_be), c("mz", "intensity"))
 })
 
 test_that("intensity<-,MsqlBackend works", {