+++ NEWS.md
@@ -1,5 +1,10 @@
 # MsqlBackend 0.0
 
+## Changes in 0.0.3
+
+- Add `peaksVariables` methods and add support for parameter `columns` in
+  `peaksData` (Requires `Spectra` version 1.15.17).
+
 ## Changes in 0.0.2
 
 - First full implementation of `MsqlBackend`.